wip: hoist Batch and BatchReader to storage-api
