+++ src/storage/src/memtable.rs
@@ -22,13 +22,13 @@ use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::Arc;
 
 use datatypes::vectors::VectorRef;
+use store_api::storage::batch::Batch;
 use store_api::storage::{consts, OpType, SequenceNumber};
 
 use crate::error::Result;
 use crate::memtable::btree::BTreeMemtable;
 pub use crate::memtable::inserter::Inserter;
 pub use crate::memtable::version::MemtableVersion;
-use crate::read::Batch;
 use crate::schema::{ProjectedSchemaRef, RegionSchemaRef};
 
 /// Unique id for memtables under same region.
